clean up styles and simplifications
