+++ frontend/src/views/OmakaseTamsPlayer/style.css
@@ -1,47 +1,31 @@
-#omakase-video-container {
-  width: 100%;
-  max-width: 100%;
-  overflow: hidden;
-}
-
 /* Force Omakase player to respect container width */
 #omakase-video-container .omakase-player,
 #omakase-video-container .omakase-player-wrapper,
 #omakase-video-container video {
-  width: 100%;
   max-width: 100%;
   max-height: 50vh;
 }
 
 #omakase-timeline {
-  width: 100%;
-  max-width: 100%;
   overflow: hidden;
-}
-
-.time-range-picker-wrapper {
-  width: 100%;
 }
 
 /* Override hardcoded widths in Omakase time range selector */
 .omakase-time-range-selector-container {
   max-width: 100% !important;
 }
-
 .omakase-time-range-selector-segments {
   width: 100% !important;
 }
 
-.omakase-player {
-  max-width: 100%;
-}
-
+/* Hide Cloudscape focus rings on library-managed elements */
 #omakase-video-container *:focus,
 #omakase-timeline *:focus {
   outline: none;
   box-shadow: none;
 }
 
+/* Allow ColumnLayout children to shrink below their content's min-width */
 #omakase-marker-list,
 #omakase-marker-toolbar {
   min-width: 0;
